+++ lib/src/duolingo_request_header.dart
@@ -38,6 +38,8 @@ class DuolingoRequestHeader {
   Map<String, String> toMap() => {
         'Cookie': cookie.rawData,
         authHeader.name: authHeader.value,
+        'Content-type': 'application/json',
+        'Accept': 'application/json',
       };
 
   @override